// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-collaboration-team-model.ts
export class RFQCollaborationTeamModel{
    id: string;
     rfqId: string;
     personName: string;
     position: string;
     userId: string;
     substituteUserId: string;
     substitutePerson: string;
     dateAndTime: string;
    rfqCollaborationTeamAccessModels: any[];
    collaborationTeamId: string;


}